MP authentication refactoring
