+++ src/main/AndroidManifest.xml
@@ -8,6 +8,6 @@
     <application android:allowBackup="true" android:label="@string/app_name"
         android:supportsRtl="true">
         <service android:name=".RadarService"/>
-        <service android:name=".auth.ManagementPortalService"/>
+        <service android:name=".auth.portal.ManagementPortalService"/>
     </application>
 </manifest>
