+++ src/main/java/org/radarcns/android/auth/portal/ManagementPortalService.java
@@ -1,4 +1,4 @@
-package org.radarcns.android.auth;
+package org.radarcns.android.auth.portal;
 
 import android.app.IntentService;
 import android.content.Context;
@@ -6,11 +6,12 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.ResultReceiver;
 import android.util.SparseArray;
-import okhttp3.Call;
-import okhttp3.Callback;
-import okhttp3.Response;
 import org.json.JSONException;
 import org.radarcns.android.RadarConfiguration;
+import org.radarcns.android.auth.AppAuthState;
+import org.radarcns.android.auth.AppSource;
+import org.radarcns.android.auth.AuthStringParser;
+import org.radarcns.android.util.SynchronousCallback;
 import org.radarcns.config.ServerConfig;
 import org.radarcns.producer.AuthenticationException;
 import org.slf4j.Logger;
@@ -26,32 +27,44 @@ import static org.radarcns.android.RadarConfiguration.MANAGEMENT_PORTAL_URL_KEY;
 import static org.radarcns.android.RadarConfiguration.OAUTH2_CLIENT_ID;
 import static org.radarcns.android.RadarConfiguration.OAUTH2_CLIENT_SECRET;
 import static org.radarcns.android.RadarConfiguration.UNSAFE_KAFKA_CONNECTION;
-import static org.radarcns.android.auth.ManagementPortalClient.SOURCES_PROPERTY;
-import static org.radarcns.android.auth.ManagementPortalClient.parseAccessToken;
+import static org.radarcns.android.auth.portal.ManagementPortalClient.MP_REFRESH_TOKEN_PROPERTY;
+import static org.radarcns.android.auth.portal.ManagementPortalClient.SOURCES_PROPERTY;
 import static org.radarcns.android.device.DeviceServiceProvider.SOURCE_KEY;
 
 /**
  * Handles intents to register sources to the ManagementPortal.
+ *
+ * It also keeps a cache of the access token, refresh token and registered sources.
  */
 public class ManagementPortalService extends IntentService {
     public static final String RESULT_RECEIVER_PROPERTY = ManagementPortalService.class.getName()
             + ".resultReceiver";
     public static final int MANAGEMENT_PORTAL_REGISTRATION = 1;
     public static final int MANAGEMENT_PORTAL_REGISTRATION_FAILED = 2;
-    public static final int MANAGEMENT_PORTAL_REFRESH = 3;
-    public static final int MANAGEMENT_PORTAL_REFRESH_FAILED = 4;
+    public static final int MANAGEMENT_PORTAL_REFRESH = 4;
+    public static final int MANAGEMENT_PORTAL_REFRESH_FAILED = 5;
 
     private static final Logger logger = LoggerFactory.getLogger(ManagementPortalService.class);
-    private static final String REGISTER_SOURCE_ACTION = "registerAction";
-    private static final String REFREST_TOKEN_ACTION = "refreshTokenAction";
+    private static final String REGISTER_SOURCE_ACTION = "org.radarcns.android.auth.ManagementPortalService.registerSourceAction";
+    private static final String GET_STATE_ACTION = "org.radarcns.android.auth.ManagementPortalService.getStateAction";
+    private static final String REFRESH_TOKEN_ACTION = "org.radarcns.android.auth.ManagementPortalService.refreshTokenAction";
+    private static final String REFRESH_TOKEN_KEY = "org.radarcns.android.auth.ManagementPortalService.refreshToken";
+    private static final String UPDATE_SUBJECT_KEY = "org.radarcns.android.auth.ManagementPortalService.updateSubject";
     private ManagementPortalClient client;
     private SparseArray<AppSource> sources;
     private String clientSecret;
     private String clientId;
+    private AppAuthState authState;
 
     public ManagementPortalService() {
         super(ManagementPortalService.class.getName());
         sources = new SparseArray<>();
+    }
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        authState = AppAuthState.Builder.from(this).build();
     }
 
     @Override
@@ -62,8 +75,11 @@ public class ManagementPortalService extends IntentService {
             case REGISTER_SOURCE_ACTION:
                 registerSource(intent);
                 break;
-            case REFREST_TOKEN_ACTION:
+            case REFRESH_TOKEN_ACTION:
                 refreshToken(intent);
+                break;
+            case GET_STATE_ACTION:
+                getState(intent);
                 break;
             default:
                 // do nothing
@@ -71,36 +87,37 @@ public class ManagementPortalService extends IntentService {
         }
     }
 
+    private void getState(Intent intent) {
+        ResultReceiver receiver = intent.getParcelableExtra(RESULT_RECEIVER_PROPERTY);
+        Bundle result = new Bundle();
+        authState.addToBundle(result);
+        receiver.send(MANAGEMENT_PORTAL_REFRESH, result);
+    }
+
     private void refreshToken(Intent intent) {
         logger.info("Refreshing JWT");
-        final ResultReceiver receiver = intent.getParcelableExtra(RESULT_RECEIVER_PROPERTY);
+        ResultReceiver receiver = intent.getParcelableExtra(RESULT_RECEIVER_PROPERTY);
         try {
-            final AppAuthState state = AppAuthState.Builder.from(intent.getExtras()).build();
-            client.refreshToken(state, clientId, clientSecret, new Callback() {
-                @Override
-                public void onFailure(Call call, IOException e) {
-                    logger.info("Failed to refresh JWT", e);
-                    receiver.send(MANAGEMENT_PORTAL_REFRESH_FAILED, null);
-                }
-
-                @Override
-                public void onResponse(Call call, Response response) {
-                    try {
-                        AppAuthState updatedState = parseAccessToken(state, response);
-                        Bundle result = new Bundle();
-                        updatedState.addToBundle(result);
-                        logger.info("Refreshed JWT");
-                        receiver.send(MANAGEMENT_PORTAL_REFRESH, result);
-                    } catch (IOException ex) {
-                        onFailure(call, ex);
-                    }
-                }
-            });
+            String refreshToken = intent.getStringExtra(REFRESH_TOKEN_KEY);
+            if (refreshToken != null) {
+                authState.newBuilder()
+                        .property(MP_REFRESH_TOKEN_PROPERTY, refreshToken)
+                        .build();
+            }
+            AuthStringParser parser;
+            if (intent.getBooleanExtra(UPDATE_SUBJECT_KEY, false)) {
+                parser = new SubjectTokenParser(client, authState);
+            } else {
+                parser = new AccessTokenParser(authState);
+            }
+            SynchronousCallback<AppAuthState> callback = new SynchronousCallback<>(parser);
+            client.refreshToken(authState, clientId, clientSecret, callback);
+            authState = callback.get();
 
             Bundle result = new Bundle();
-            state.addToBundle(result);
+            authState.addToBundle(result);
             receiver.send(MANAGEMENT_PORTAL_REFRESH, result);
-        } catch (JSONException | IOException e) {
+        } catch (JSONException | IOException | InterruptedException e) {
             receiver.send(MANAGEMENT_PORTAL_REFRESH_FAILED, null);
         }
     }
@@ -110,7 +127,6 @@ public class ManagementPortalService extends IntentService {
         AppSource source = intent.getParcelableExtra(SOURCE_KEY);
 
         AppSource resultSource = sources.get((int)source.getSourceTypeId());
-        AppAuthState authState = AppAuthState.Builder.from(intent.getExtras()).build();
         ResultReceiver receiver = intent.getParcelableExtra(RESULT_RECEIVER_PROPERTY);
 
         if (resultSource == null) {
@@ -120,10 +136,11 @@ public class ManagementPortalService extends IntentService {
                 @SuppressWarnings("unchecked")
                 List<AppSource> existingSources = (List<AppSource>) authState.getProperties().get(SOURCES_PROPERTY);
 
+                boolean containsSource = false;
+                ArrayList<AppSource> updatedSources;
                 if (existingSources != null) {
-                    ArrayList<AppSource> updatedSources = new ArrayList<>(existingSources.size());
+                    updatedSources = new ArrayList<>(existingSources.size());
                     updatedSources.add(resultSource);
-                    boolean containsSource = false;
                     for (AppSource existingSource : existingSources) {
                         if (existingSource.getSourceTypeId() != resultSource.getSourceTypeId()) {
                             updatedSources.add(existingSource);
@@ -131,30 +148,31 @@ public class ManagementPortalService extends IntentService {
                             containsSource = true;
                         }
                     }
-                    authState = authState.newBuilder()
-                            .property(SOURCES_PROPERTY, updatedSources)
-                            .build();
-                    if (!containsSource) {
-                        authState.invalidate(getApplicationContext());
-                    }
+                } else {
+                    updatedSources = new ArrayList<>(1);
+                    updatedSources.add(resultSource);
                 }
+                authState = authState.newBuilder()
+                        .property(SOURCES_PROPERTY, updatedSources)
+                        .build();
+
+                if (!containsSource) {
+                    authState.invalidate(getApplicationContext());
+                }
+
+                Bundle result = new Bundle();
+                authState.addToBundle(result);
+                result.putParcelable(SOURCE_KEY, resultSource);
+                receiver.send(MANAGEMENT_PORTAL_REGISTRATION, result);
             } catch (AuthenticationException ex) {
                 authState.invalidate(getApplicationContext());
             } catch (IOException | JSONException ex) {
                 logger.error("Failed to register source {} of type {} {}",
                         source.getSourceName(), source.getSourceTypeProducer(),
                         source.getSourceTypeModel(), ex);
-                if (receiver != null) {
-                    receiver.send(MANAGEMENT_PORTAL_REGISTRATION_FAILED, null);
-                }
-                return;
+                receiver.send(MANAGEMENT_PORTAL_REGISTRATION_FAILED, null);
             }
         }
-
-        Bundle result = new Bundle();
-        result.putParcelable(SOURCE_KEY, resultSource);
-        authState.addToBundle(result);
-        receiver.send(MANAGEMENT_PORTAL_REGISTRATION, result);
     }
 
     private void ensureClient() {
@@ -178,28 +196,26 @@ public class ManagementPortalService extends IntentService {
     }
 
     /** Build an intent to create a request for the management portal. */
-    public static Intent createRequest(Context context,
-            AppSource source, AppAuthState state, ResultReceiver receiver) {
+    public static void registerSource(Context context, AppSource source, ResultReceiver receiver) {
         Intent intent = new Intent(context, ManagementPortalService.class);
         intent.setAction(REGISTER_SOURCE_ACTION);
         Bundle extras = new Bundle();
         extras.putParcelable(SOURCE_KEY, source);
-        state.addToBundle(extras);
         extras.putParcelable(RESULT_RECEIVER_PROPERTY, receiver);
         intent.putExtras(extras);
-        return intent;
+        context.startService(intent);
     }
 
-
     /** Build an intent to create a request for the management portal. */
-    public static Intent createRequest(Context context, AppAuthState state, ResultReceiver receiver) {
+    public static void requestAccessToken(Context context, String refreshToken, boolean updateSubject, ResultReceiver receiver) {
         Intent intent = new Intent(context, ManagementPortalService.class);
-        intent.setAction(REFREST_TOKEN_ACTION);
+        intent.setAction(REFRESH_TOKEN_ACTION);
         Bundle extras = new Bundle();
-        state.addToBundle(extras);
+        extras.putString(REFRESH_TOKEN_KEY, refreshToken);
         extras.putParcelable(RESULT_RECEIVER_PROPERTY, receiver);
+        extras.putBoolean(UPDATE_SUBJECT_KEY, updateSubject);
         intent.putExtras(extras);
-        return intent;
+        context.startService(intent);
     }
 
     public static boolean isEnabled() {
@@ -208,6 +224,7 @@ public class ManagementPortalService extends IntentService {
 
     @Override
     public void onDestroy() {
+        authState.addToPreferences(this);
         if (client != null) {
             client.close();
         }
