+++ src/main/java/org/radarcns/android/util/Parser.java
@@ -1,0 +1,9 @@
+package org.radarcns.android.util;
+
+import android.support.annotation.NonNull;
+
+import java.io.IOException;
+
+public interface Parser<S, T> {
+    T parse(@NonNull S type) throws IOException;
+}
